+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Repository title
 - `from_globals()`
+- `Router::getAllowedMethods()`
 
 ### Changed
 - Use `class` keyword in foreign classes